4Z33DKP split IssueCommits.tsx: diff-selection and row styles to lib, DiffStat, SelectionComposer, FileRow and CommitRow to their own files
