+++ source/gui/client/components/IssueCommits.tsx
@@ -1,39 +1,17 @@
 import React, {useEffect, useRef, useState} from 'react';
 import {
-	DiffLineAnnotation,
-	SelectedLineRange,
-	SelectionSide,
-} from '@pierre/diffs/react';
-import {
 	GuiComment,
 	GuiCommitDiffFile,
 	GuiRefCommitEntry,
 } from '../lib/gui-state.model';
-import {CONTENT_FONT, GUI_THEME, TEXT} from '../lib/gui-theme';
-import {diffLineCount, isLargeDiff} from '../../../lib/utils/diff-size.js';
-import {COMMENT_CARD_STYLE} from '../lib/comment-card.style';
-import {
-	DiffCommentMeta,
-	encodeDiffCommentMarker,
-	extractCommentSnippet,
-	formatSelectionLabel,
-	isSelectionSide,
-	parseDiffCommentMeta,
-	stripDiffCommentMarker,
-} from '../../../lib/utils/diff-comment.js';
-import {timeAgo} from '../lib/gui-format.helper';
-import {ReviewedFiles, useReviewedFiles} from '../lib/reviewed-files';
-import {ActionRow, Textarea} from './FormPrimitives';
-import {Button} from './Button';
-import {Checkbox} from './Checkbox';
+import {GUI_THEME} from '../lib/gui-theme';
+import {isLargeDiff} from '../../../lib/utils/diff-size.js';
+import {useReviewedFiles} from '../lib/reviewed-files';
 import {CopyRef} from './CopyRef';
-import {CreateNodeModal} from './CreateNodeModal';
-import {CopyShaButton} from './CopyShaButton';
 import {Empty} from './FormPrimitives';
-import {DIFF_BOX_STYLE, FileDiffView} from './DiffPanel';
-import {IconChevronDown} from './IconChevronDown';
-import {IconChevronRight} from './IconChevronRight';
-import {IconComment} from './IconComment';
+import {FileTicketParams, CommitFocus} from '../lib/diff-selection';
+import {RAIL_WIDTH, RAIL_DOT_OFFSET, ROW_GAP} from '../lib/commit-row.style';
+import {CommitRow} from './CommitRow';
 
 export type CommitDiffState = {
 	loading: boolean;
@@ -50,907 +28,6 @@ const stripRefPrefix = (subject: string, ref: string): string => {
 	return subject.toUpperCase().startsWith(prefix.toUpperCase())
 		? subject.slice(prefix.length)
 		: subject;
-};
-
-// A selection's start/end are the real (gutter-displayed) line numbers within
-// whichever side they belong to — 'deletions' means the old file, 'additions'
-// the new one. A range can span both sides (dragged from a removed line into
-// an added one in split view): quote both halves rather than picking one.
-export const extractSnippet = (
-	file: GuiCommitDiffFile,
-	range: SelectedLineRange,
-): string => {
-	const linesFor = (side: SelectedLineRange['side']) =>
-		(side === 'deletions' ? file.before : file.after).split('\n');
-
-	const endSide = range.endSide ?? range.side;
-
-	if (endSide === range.side) {
-		return linesFor(range.side)
-			.slice(range.start - 1, range.end)
-			.join('\n');
-	}
-
-	const startHalf = linesFor(range.side).slice(range.start - 1);
-	const endHalf = linesFor(endSide).slice(0, range.end);
-
-	return [...startHalf, ...endHalf].join('\n');
-};
-
-export {formatSelectionLabel} from '../../../lib/utils/diff-comment.js';
-
-// Quoted lines keep their real source indentation (often several tabs deep
-// inside nested JSX) — fine in the wide diff, unreadable in a narrow comment
-// box. Strips the whitespace every non-blank line shares, same as most
-// editors' own "copy" behavior.
-export const dedent = (snippet: string): string => {
-	const lines = snippet.split('\n');
-
-	const commonIndent = lines
-		.filter(line => line.trim() !== '')
-		.reduce<number | null>((min, line) => {
-			const indent = /^[ \t]*/.exec(line)?.[0].length ?? 0;
-			return min === null ? indent : Math.min(min, indent);
-		}, null);
-
-	if (!commonIndent) return snippet;
-
-	return lines.map(line => line.slice(commonIndent)).join('\n');
-};
-
-// The marker codec lives in lib so the TUI renders the same bodies; the GUI
-// keeps reaching it through this module.
-export {
-	encodeDiffCommentMarker,
-	extractCommentLead,
-	extractCommentSnippet,
-	parseDiffCommentMeta,
-	stripDiffCommentMarker,
-	withDiffCommentNote,
-} from '../../../lib/utils/diff-comment.js';
-export type {DiffCommentMeta} from '../../../lib/utils/diff-comment.js';
-
-export type DiffComment = {comment: GuiComment; meta: DiffCommentMeta};
-
-export const findDiffCommentsForFile = (
-	comments: GuiComment[],
-	filePath: string,
-): DiffComment[] =>
-	comments.flatMap(comment => {
-		if (comment.isDeleted) return [];
-		const meta = parseDiffCommentMeta(comment.body);
-		return meta && meta.filePath === filePath ? [{comment, meta}] : [];
-	});
-
-// What a "File ticket" submission carries up to the caller that owns the
-// actual issues:create call and the origin-ticket back-comment — everything
-// needed to build both without the caller re-deriving any of it.
-export type FileTicketParams = {
-	sha: string;
-	filePath: string;
-	range: SelectedLineRange;
-	snippet: string;
-	title: string;
-	note: string;
-};
-
-// A spot in a ticket's Commits tab, deep-linkable from a comment. Lives in the
-// URL rather than in transient state so the link survives a reload and can be
-// handed to someone else.
-export type DiffLocation = {
-	sha: string;
-	filePath: string;
-	start: number;
-	end: number;
-	side: SelectionSide;
-	endSide: SelectionSide;
-	// Set when the diff lives on another ticket's Commits tab.
-	issueRef?: string;
-};
-
-export const diffLocationFromMeta = (
-	meta: DiffCommentMeta,
-): DiffLocation | null =>
-	meta.sha
-		? {
-				sha: meta.sha,
-				filePath: meta.filePath,
-				start: meta.start,
-				end: meta.end,
-				side: meta.side,
-				endSide: meta.endSide,
-				...(meta.issueRef ? {issueRef: meta.issueRef} : {}),
-		  }
-		: null;
-
-const DIFF_LOCATION_PARAMS = [
-	'commit',
-	'file',
-	'from',
-	'to',
-	'side',
-	'endSide',
-] as const;
-
-export const writeDiffLocationParams = (
-	params: URLSearchParams,
-	location: DiffLocation,
-): void => {
-	params.set('commit', location.sha);
-	params.set('file', location.filePath);
-	params.set('from', String(location.start));
-	params.set('to', String(location.end));
-	params.set('side', location.side);
-	params.set('endSide', location.endSide);
-};
-
-// A deep link that names only a commit: open it, nothing narrower to show.
-// A full DiffLocation is one of these too.
-export type CommitFocus = {sha: string} & Partial<Omit<DiffLocation, 'sha'>>;
-
-export const readCommitFocusParam = (
-	params: URLSearchParams,
-): CommitFocus | null => {
-	const sha = params.get('commit');
-	return sha ? {sha} : null;
-};
-
-export const clearDiffLocationParams = (params: URLSearchParams): void => {
-	for (const key of DIFF_LOCATION_PARAMS) params.delete(key);
-};
-
-export const readDiffLocationParams = (
-	params: URLSearchParams,
-): DiffLocation | null => {
-	const sha = params.get('commit');
-	const filePath = params.get('file');
-	const start = Number(params.get('from'));
-	const end = Number(params.get('to'));
-	const side = params.get('side');
-	const endSide = params.get('endSide');
-
-	if (
-		!sha ||
-		!filePath ||
-		!Number.isFinite(start) ||
-		!Number.isFinite(end) ||
-		!isSelectionSide(side) ||
-		!isSelectionSide(endSide)
-	) {
-		return null;
-	}
-
-	return {sha, filePath, start, end, side, endSide};
-};
-
-// The timeline rail: a dot per commit, in the scrubber's own commit-series
-// color, connected to the next by a line. RAIL_DOT_OFFSET lines the dot up
-// with the header's text — the card's border, then the header's padding,
-// then half its content height (the copy button, its tallest child, at 20px,
-// with the subject centred against it) — not the row's overall height, which
-// grows when a commit is expanded.
-const RAIL_WIDTH = 24;
-const COMMIT_HEADER_PADDING = 13;
-const RAIL_DOT_OFFSET = 1 + COMMIT_HEADER_PADDING + 10;
-const ROW_GAP = 14;
-
-const disclosureStyle: React.CSSProperties = {
-	display: 'flex',
-	alignItems: 'center',
-	gap: 8,
-	width: '100%',
-	textAlign: 'left',
-	background: 'transparent',
-	border: 'none',
-	borderRadius: 6,
-	cursor: 'pointer',
-	color: GUI_THEME.primary,
-	font: 'inherit',
-	fontSize: TEXT.ui,
-	transition: 'background 120ms ease',
-};
-
-// Says "this opens" before it is clicked.
-const DISCLOSURE_HOVER_BG = 'rgba(255,255,255,0.04)';
-
-// A rounded pill rather than GitHub's five solid squares — matches the
-// rest of the app's soft, rounded chrome instead of copying its exact look.
-const DiffStat = ({
-	insertions,
-	deletions,
-}: {
-	insertions: number;
-	deletions: number;
-}) => {
-	const total = insertions + deletions;
-	if (total === 0) return null;
-
-	const addRatio = insertions / total;
-
-	return (
-		<div
-			style={{
-				display: 'flex',
-				alignItems: 'center',
-				gap: 6,
-				flexShrink: 0,
-				fontFamily: 'ui-monospace, monospace',
-				fontSize: TEXT.meta,
-			}}
-		>
-			<span style={{color: GUI_THEME.green}}>+{insertions}</span>
-			<span style={{color: GUI_THEME.red}}>-{deletions}</span>
-			<div
-				style={{
-					width: 32,
-					height: 3,
-					borderRadius: 1.5,
-					overflow: 'hidden',
-					display: 'flex',
-					background: GUI_THEME.line,
-				}}
-			>
-				<div
-					style={{width: `${addRatio * 100}%`, background: GUI_THEME.green}}
-				/>
-				<div
-					style={{
-						width: `${(1 - addRatio) * 100}%`,
-						background: GUI_THEME.red,
-					}}
-				/>
-			</div>
-		</div>
-	);
-};
-
-// Rendered by MultiFileDiff right under the last selected line. One box for
-// both outcomes: write the note first, then choose whether it becomes a
-// comment on this ticket or a new ticket of its own.
-const SelectionComposer = ({
-	sha,
-	file,
-	selection,
-	note,
-	onChangeNote,
-	onAddComment,
-	onFileTicket,
-	onClear,
-}: {
-	sha: string;
-	file: GuiCommitDiffFile;
-	selection: SelectedLineRange;
-	note: string;
-	onChangeNote: (note: string) => void;
-	onAddComment?: (body: string) => void;
-	// Opens the title prompt; the row owns the actual filing.
-	onFileTicket?: () => void;
-	onClear: () => void;
-}) => {
-	const snippet = dedent(extractSnippet(file, selection));
-	const selectionLabel = `${file.path} ${formatSelectionLabel(selection)}`;
-
-	const comment = () => {
-		const trimmedNote = note.trim();
-		// Matches extractSnippet's own default: a range without a reported side
-		// is expected to be a modern-git edge case at worst, not a real gap.
-		const side = selection.side ?? 'additions';
-		const endSide = selection.endSide ?? side;
-
-		const marker = encodeDiffCommentMarker({
-			filePath: file.path,
-			start: selection.start,
-			side,
-			end: selection.end,
-			endSide,
-			note: trimmedNote,
-			sha,
-		});
-
-		const body = [
-			...(trimmedNote ? [trimmedNote, ''] : []),
-			marker,
-			`\`${file.path}\` ${formatSelectionLabel(selection)}`,
-			'```',
-			snippet,
-			'```',
-		].join('\n');
-
-		onAddComment?.(body);
-		onClear();
-	};
-
-	return (
-		<div
-			data-testid="selection-composer"
-			style={{
-				margin: '4px 0',
-				padding: 10,
-				border: `1px solid ${GUI_THEME.accent}`,
-				borderRadius: 8,
-				background: GUI_THEME.tertiary,
-				boxShadow: `0 0 0 1px ${GUI_THEME.accent}33`,
-			}}
-		>
-			<div style={{fontSize: TEXT.meta, color: GUI_THEME.secondary}}>
-				{selectionLabel}
-			</div>
-
-			<Textarea
-				autoFocus
-				maxLength={Number.MAX_SAFE_INTEGER}
-				value={note}
-				placeholder="Add a note (optional)"
-				onChange={event => onChangeNote(event.target.value)}
-				onKeyDown={event => {
-					if (event.key === 'Escape') onClear();
-					if ((event.metaKey || event.ctrlKey) && event.key === 'Enter') {
-						if (onAddComment) comment();
-					}
-				}}
-				style={{
-					marginTop: 8,
-					minHeight: 45,
-					font: 'inherit',
-					fontFamily: CONTENT_FONT,
-					fontSize: TEXT.prose,
-				}}
-			/>
-
-			<ActionRow>
-				<Button variant="ghost" onClick={onClear}>
-					Cancel
-				</Button>
-				{onFileTicket && (
-					<Button variant="default" onClick={onFileTicket}>
-						File ticket
-					</Button>
-				)}
-				{onAddComment && (
-					<Button variant="primary" onClick={comment}>
-						Comment
-					</Button>
-				)}
-			</ActionRow>
-		</div>
-	);
-};
-
-// What a file's diff carries at a line: an existing comment, or the composer
-// for the range being selected right now.
-type RowAnnotation =
-	| {kind: 'comment'; entry: DiffComment}
-	| {kind: 'composer'; selection: SelectedLineRange};
-
-// Rendered by MultiFileDiff at the line a diff-selection comment is anchored
-// to. Just the author and note — the full body (including the requoted
-// snippet, redundant here since the diff itself is right above it) lives in
-// the Comments tab.
-const DiffCommentAnnotation = ({
-	entry,
-	onHover,
-}: {
-	entry: DiffComment;
-	onHover: (hovering: boolean) => void;
-}) => {
-	const {comment, meta} = entry;
-
-	return (
-		<div
-			data-testid="diff-comment"
-			onMouseEnter={() => onHover(true)}
-			onMouseLeave={() => onHover(false)}
-			style={{
-				...COMMENT_CARD_STYLE,
-				margin: '4px 0',
-				fontSize: TEXT.ui,
-			}}
-		>
-			<div
-				style={{
-					display: 'flex',
-					alignItems: 'center',
-					gap: 8,
-					color: GUI_THEME.secondary,
-					fontSize: TEXT.meta,
-				}}
-			>
-				<span style={{display: 'inline-flex', color: GUI_THEME.accent}}>
-					<IconComment size={12} />
-				</span>
-				<span>
-					{comment.author.name ?? 'unknown'}
-					{comment.createdAt && (
-						<span style={{color: GUI_THEME.dim2}}>
-							{' '}
-							· {timeAgo(comment.createdAt)}
-						</span>
-					)}
-				</span>
-			</div>
-			<div style={{marginTop: 4}}>
-				{meta.note || <em style={{color: GUI_THEME.dim}}>commented</em>}
-			</div>
-		</div>
-	);
-};
-
-// The one header a file gets, open or shut. Open, the highlighter renders it
-// in its sticky slot in place of its own (change icon, name, counts), so the
-// name that stays in view while the diff scrolls is also the one that folds
-// it; shut, FileRow stands it in the same box by itself.
-const FileHeader = ({
-	file,
-	expanded,
-	onToggle,
-	commentCount,
-	reviewed,
-	onReviewed,
-}: {
-	file: GuiCommitDiffFile;
-	expanded: boolean;
-	onToggle: () => void;
-	commentCount: number;
-	reviewed: boolean;
-	onReviewed: (next: boolean) => void;
-}) => {
-	const [lit, setLit] = useState(false);
-
-	return (
-		<div
-			style={{
-				display: 'flex',
-				alignItems: 'center',
-				gap: 10,
-				padding: '4px 10px 4px 4px',
-			}}
-		>
-			<button
-				onClick={onToggle}
-				aria-expanded={expanded}
-				onMouseEnter={() => setLit(true)}
-				onMouseLeave={() => setLit(false)}
-				onFocus={() => setLit(true)}
-				onBlur={() => setLit(false)}
-				style={{
-					...disclosureStyle,
-					flex: 1,
-					minWidth: 0,
-					width: 'auto',
-					color: GUI_THEME.secondary,
-					padding: '4px 6px',
-					background: lit ? DISCLOSURE_HOVER_BG : 'transparent',
-				}}
-			>
-				{expanded ? (
-					<IconChevronDown size={12} />
-				) : (
-					<IconChevronRight size={12} />
-				)}
-				<span
-					style={{
-						fontFamily: 'ui-monospace, monospace',
-						overflow: 'hidden',
-						textOverflow: 'ellipsis',
-						whiteSpace: 'nowrap',
-					}}
-				>
-					{file.path}
-				</span>
-				{/* Findable when collapsed — otherwise a comment left on a file with
-				    several others is easy to lose track of. */}
-				{commentCount > 0 && (
-					<span
-						title={`${commentCount} comment${commentCount === 1 ? '' : 's'}`}
-						style={{
-							display: 'inline-flex',
-							alignItems: 'center',
-							gap: 3,
-							flexShrink: 0,
-							color: GUI_THEME.accent,
-							fontSize: TEXT.label,
-						}}
-					>
-						<IconComment size={10} />
-						{commentCount}
-					</span>
-				)}
-				{/* Says why "Expand all" passed this one over. Only while shut: once
-				    it is open the diff speaks for itself. */}
-				{!expanded && isLargeDiff(file) && (
-					<span
-						data-testid="large-diff-badge"
-						title={`${diffLineCount(
-							file,
-						).toLocaleString()} lines — left collapsed, open it to load the diff`}
-						style={{
-							flexShrink: 0,
-							color: GUI_THEME.dim,
-							fontSize: TEXT.label,
-							whiteSpace: 'nowrap',
-						}}
-					>
-						large diff
-					</span>
-				)}
-			</button>
-			<DiffStat insertions={file.insertions} deletions={file.deletions} />
-			{/* Beside the toggle rather than inside it: ticking a file off must
-			    not also fold or unfold it by accident. */}
-			<Checkbox
-				label="reviewed"
-				checked={reviewed}
-				onChange={onReviewed}
-				activeColor={GUI_THEME.green}
-			/>
-		</div>
-	);
-};
-
-const FileRow = ({
-	sha,
-	file,
-	expanded,
-	onToggle,
-	diffStyle,
-	onAddComment,
-	onFileTicket,
-	comments,
-	focusRange,
-	reviewed,
-	onReviewed,
-}: {
-	sha: string;
-	file: GuiCommitDiffFile;
-	expanded: boolean;
-	onToggle: () => void;
-	reviewed: boolean;
-	onReviewed: (next: boolean) => void;
-	diffStyle: 'split' | 'unified';
-	onAddComment?: (body: string) => void;
-	onFileTicket?: (params: FileTicketParams) => void;
-	comments: GuiComment[];
-	// Set when a comment permalink points at this file: seeds the highlight
-	// and scrolls the diff into view.
-	focusRange?: SelectedLineRange | null;
-}) => {
-	const [selection, setSelection] = useState<SelectedLineRange | null>(null);
-	// Held here rather than in the composer so re-dragging the range keeps
-	// what was typed.
-	const [note, setNote] = useState('');
-	const rowRef = useRef<HTMLDivElement | null>(null);
-
-	// The title being typed for a ticket filed off the selection; null while
-	// the prompt is closed. Asked for rather than inferred from the note, so
-	// the note stays a note.
-	const [ticketTitle, setTicketTitle] = useState<string | null>(null);
-	// The range of the comment under the pointer, lit up so the reader can
-	// see which lines it is about.
-	const [hoveredRange, setHoveredRange] = useState<SelectedLineRange | null>(
-		null,
-	);
-
-	const clearSelection = () => {
-		setSelection(null);
-		setNote('');
-		setTicketTitle(null);
-	};
-
-	const fileTicket = () => {
-		const title = ticketTitle?.trim();
-		if (!title || !selection) return;
-
-		onFileTicket?.({
-			sha,
-			filePath: file.path,
-			range: selection,
-			snippet: dedent(extractSnippet(file, selection)),
-			title,
-			note: note.trim(),
-		});
-		clearSelection();
-	};
-
-	// Keyed on the range's own values, not object identity — the parent
-	// rebuilds it from URL params on every render, and depending on identity
-	// would re-scroll (and re-select) forever.
-	const focusKey = focusRange
-		? `${focusRange.start}-${focusRange.end}-${focusRange.side}-${focusRange.endSide}`
-		: null;
-
-	useEffect(() => {
-		if (!focusRange || !expanded) return;
-
-		setSelection(focusRange);
-		rowRef.current?.scrollIntoView({block: 'center', behavior: 'smooth'});
-	}, [focusKey, expanded]);
-
-	const fileComments = findDiffCommentsForFile(comments, file.path);
-
-	const lineAnnotations: DiffLineAnnotation<RowAnnotation>[] = fileComments.map(
-		entry => ({
-			side: entry.meta.endSide,
-			lineNumber: entry.meta.end,
-			metadata: {kind: 'comment', entry},
-		}),
-	);
-
-	if (selection) {
-		const side = selection.endSide ?? selection.side ?? 'additions';
-
-		lineAnnotations.push({
-			side,
-			lineNumber: selection.end,
-			metadata: {kind: 'composer', selection},
-		});
-	}
-
-	const header = () => (
-		<FileHeader
-			file={file}
-			expanded={expanded}
-			onToggle={onToggle}
-			commentCount={fileComments.length}
-			reviewed={reviewed}
-			onReviewed={onReviewed}
-		/>
-	);
-
-	return (
-		<div ref={rowRef} data-testid="file-row">
-			{expanded ? (
-				<>
-					<FileDiffView
-						file={file}
-						diffStyle={diffStyle}
-						selectedLines={hoveredRange ?? selection}
-						onSelectionEnd={setSelection}
-						lineAnnotations={lineAnnotations}
-						renderCustomHeader={header}
-						renderAnnotation={({metadata}) =>
-							metadata.kind === 'comment' ? (
-								<DiffCommentAnnotation
-									entry={metadata.entry}
-									onHover={hovering => {
-										const {start, end, side, endSide} = metadata.entry.meta;
-										setHoveredRange(
-											hovering ? {start, end, side, endSide} : null,
-										);
-									}}
-								/>
-							) : (
-								<SelectionComposer
-									sha={sha}
-									file={file}
-									selection={metadata.selection}
-									note={note}
-									onChangeNote={setNote}
-									onAddComment={onAddComment}
-									// Prefilled with the note's first line: what was just
-									// written is usually the title, and is still editable.
-									onFileTicket={
-										onFileTicket
-											? () => setTicketTitle(note.trim().split('\n')[0] ?? '')
-											: undefined
-									}
-									onClear={clearSelection}
-								/>
-							)
-						}
-					/>
-					{ticketTitle !== null && selection && (
-						<CreateNodeModal
-							eyebrow="File a ticket"
-							fieldLabel={`${file.path} ${formatSelectionLabel(selection)}`}
-							placeholder="Ticket title"
-							confirmLabel="file ticket"
-							title={ticketTitle}
-							onChangeTitle={setTicketTitle}
-							onCreate={fileTicket}
-							onClose={() => setTicketTitle(null)}
-						/>
-					)}
-				</>
-			) : (
-				// No diff mounted while shut — a large file left collapsed is
-				// meant to cost nothing — so the header stands in the same box
-				// by itself.
-				<div style={DIFF_BOX_STYLE}>{header()}</div>
-			)}
-		</div>
-	);
-};
-
-const CommitRow = ({
-	commit,
-	diff,
-	expanded,
-	onToggle,
-	expandedFiles,
-	onToggleFile,
-	onSetAllFilesExpanded,
-	reviewedFiles,
-	diffStyle,
-	onAddComment,
-	onFileTicket,
-	comments,
-	focus,
-}: {
-	commit: GuiRefCommitEntry;
-	diff: CommitDiffState | undefined;
-	expanded: boolean;
-	onToggle: () => void;
-	expandedFiles: Set<string>;
-	onToggleFile: (path: string) => void;
-	onSetAllFilesExpanded: (filePaths: string[], expand: boolean) => void;
-	reviewedFiles: ReviewedFiles;
-	diffStyle: 'split' | 'unified';
-	onAddComment?: (body: string) => void;
-	onFileTicket?: (params: FileTicketParams) => void;
-	comments: GuiComment[];
-	// Non-null only on the commit a permalink points at.
-	focus?: CommitFocus | null;
-}) => {
-	const [hovered, setHovered] = useState(false);
-	// Also tracks focus (not just mouse hover): the copy button is a real
-	// nested <button>, so a keyboard user can Tab straight to it — without
-	// this it would sit there focused but invisible (opacity 0) until Enter.
-	const [focused, setFocused] = useState(false);
-	const revealed = hovered || focused;
-
-	// Large files are not part of "expand all", so the button reads its state
-	// off the ones that are. The fallback covers a commit where every file is
-	// large, so the button can still collapse what was opened by hand rather
-	// than sitting there doing nothing.
-	const expandablePaths = (diff?.files ?? [])
-		.filter(file => !isLargeDiff(file))
-		.map(file => file.path);
-	const allExpandableExpanded =
-		expandablePaths.length > 0
-			? expandablePaths.every(path => expandedFiles.has(path))
-			: (diff?.files ?? []).some(file => expandedFiles.has(file.path));
-
-	return (
-		<div
-			data-testid="commit-card"
-			style={{
-				border: `1px solid ${GUI_THEME.line}`,
-				borderRadius: 8,
-				// Clip rather than hide, so the card rounds its corners without
-				// becoming a scroll container — one would trap each file's sticky
-				// header in a box that never scrolls.
-				overflow: 'clip',
-			}}
-		>
-			{/* A div, not a button: it holds CopyShaButton, a real nested button,
-		    which native <button> nesting forbids. role/tabIndex/onKeyDown stand
-		    in for what the element would otherwise give for free. */}
-			<div
-				role="button"
-				tabIndex={0}
-				onClick={onToggle}
-				onKeyDown={event => {
-					if (event.key === 'Enter' || event.key === ' ') {
-						event.preventDefault();
-						onToggle();
-					}
-				}}
-				onMouseEnter={() => setHovered(true)}
-				onMouseLeave={() => setHovered(false)}
-				onFocus={() => setFocused(true)}
-				onBlur={() => setFocused(false)}
-				aria-expanded={expanded}
-				style={{
-					...disclosureStyle,
-					padding: `${COMMIT_HEADER_PADDING}px 14px`,
-					background: revealed ? DISCLOSURE_HOVER_BG : 'transparent',
-				}}
-			>
-				{/* The caret leads, as it does on the file rows, so a row reads as
-			    foldable and its state is plain. flexShrink: 0 because a tight
-			    panel width would otherwise squeeze it toward invisible rather
-			    than truncate the (already-shrinkable) subject further. */}
-				<span style={{flexShrink: 0, display: 'flex'}}>
-					{expanded ? (
-						<IconChevronDown size={12} />
-					) : (
-						<IconChevronRight size={12} />
-					)}
-				</span>
-				<span
-					style={{
-						flex: 1,
-						minWidth: 0,
-						overflow: 'hidden',
-						textOverflow: 'ellipsis',
-						whiteSpace: 'nowrap',
-					}}
-				>
-					{commit.subject}
-				</span>
-				<DiffStat insertions={commit.insertions} deletions={commit.deletions} />
-				{/* The sha is lookup chrome, not part of reading the list, so it
-			    sits at the far right and only appears on hover or once expanded —
-			    reachable without a full expand, but not permanent clutter.
-			    Always mounted, opacity-toggled rather than conditionally rendered:
-			    swapping it in/out of the tree changed the row's flex height on
-			    hover, which the rail lines (sized off that height) visibly jumped
-			    with. Keeping it in flow always reserves the same space. */}
-				<span
-					style={{
-						flexShrink: 0,
-						opacity: expanded || revealed ? 1 : 0,
-						pointerEvents: expanded || revealed ? 'auto' : 'none',
-						transition: 'opacity 120ms ease',
-					}}
-				>
-					<CopyShaButton sha={commit.sha} />
-				</span>
-			</div>
-
-			{expanded && (
-				<div
-					style={{
-						padding: '0 10px 10px',
-						borderTop: `1px solid ${GUI_THEME.line}`,
-					}}
-				>
-					{diff?.loading && <Empty>Loading files…</Empty>}
-					{!diff?.loading && diff?.error && <Empty>{diff.error}</Empty>}
-
-					{diff?.files && diff.files.length > 1 && (
-						<div style={{display: 'flex', justifyContent: 'flex-end'}}>
-							<Button
-								variant="ghost"
-								onClick={() => {
-									// Expanding skips the large ones; collapsing still takes
-									// everything, including a large file opened by hand.
-									onSetAllFilesExpanded(
-										allExpandableExpanded ? [] : expandablePaths,
-										!allExpandableExpanded,
-									);
-								}}
-							>
-								{allExpandableExpanded ? 'Collapse all' : 'Expand all'}
-							</Button>
-						</div>
-					)}
-
-					{diff?.files?.map(file => (
-						<FileRow
-							key={file.path}
-							sha={commit.sha}
-							file={file}
-							expanded={expandedFiles.has(file.path)}
-							onToggle={() => onToggleFile(file.path)}
-							reviewed={reviewedFiles.isReviewed(commit.sha, file.path)}
-							onReviewed={next =>
-								reviewedFiles.setReviewed(commit.sha, file.path, next)
-							}
-							diffStyle={diffStyle}
-							onAddComment={onAddComment}
-							onFileTicket={onFileTicket}
-							comments={comments}
-							focusRange={
-								focus?.filePath === file.path &&
-								focus.start !== undefined &&
-								focus.end !== undefined &&
-								focus.side &&
-								focus.endSide
-									? {
-											start: focus.start,
-											end: focus.end,
-											side: focus.side,
-											endSide: focus.endSide,
-									  }
-									: null
-							}
-						/>
-					))}
-				</div>
-			)}
-		</div>
-	);
 };
 
 // The tree this renders:
